move HeartbeatStream out of opt
